Changes after peer review
