+++ sandbox/src/routes/r4/services/mockResponseProvider.js
@@ -1,30 +1,4 @@
-const fs = require('fs')
-const lodash = require('lodash')
-
-function mapExampleResponse(request, exampleResponseMap) {
-
-  if (request && request.payload) {
-    for (const [requestBodyPath, response] of Object.entries(exampleResponseMap)) {
-      try {
-        const exampleRequestBody = JSON.parse(fs.readFileSync(requestBodyPath))
-        var requestBody = request.payload;
-        if ('object' != (typeof requestBody)) {
-          requestBody = JSON.parse(request.payload)
-        }
-
-        if (lodash.isEqual(requestBody, exampleRequestBody)) {
-          return response;
-        }
-      } catch (err) {
-        console.error(err)
-        throw err
-      }
-    }
-  }
-
-  return null;
-}
-
+const mockResponseUtils = require('../../mockResponseUtils')
 
 module.exports = {
 
@@ -94,7 +68,7 @@ module.exports = {
       var responseMap = {
         'src/mocks/r4/requestUploadUri/requests/RequestExample.json': 'r4/requestUploadUri/responses/ResponseExample.json'
       }
-      return mapExampleResponse(request, responseMap)
+      return mockResponseUtils.mapExampleResponse(request, responseMap)
     }
   }
 }
